board delete works
