+++ src/javascripts/components/singleBoard/boardsBtn.js
@@ -1,5 +1,7 @@
 import $ from 'jquery';
 import boards from '../boards/boards';
+// import smash from '../../helpers/data/smash';
+// import singleBoard from './singleBoard';
 
 const showUserBoards = () => {
   $('#singleBoard').html('');
