+++ src/javascripts/components/singleBoard/singleBoard.scss
@@ -6,4 +6,9 @@
       border: 3px solid rgba(54, 62, 85, 0.767);
     }
   }
+
+  .del-brd{
+    max-height: 2em;
+    max-width: 2em;
+  }
 }
