+++ src/javascripts/helpers/data/boardsData.js
@@ -6,6 +6,8 @@ import apiKey from '../apiKeys.json';
 const baseUrl = apiKey.firebaseConfig.databaseURL;
 const getCurrentUid = () => firebase.auth().currentUser.uid;
 
+const deleteBoard = (boardId) => axios.delete(`${baseUrl}/boards/${boardId}.json`);
+
 const getAllBoards = () => new Promise((resolve, reject) => {
   const currentUid = getCurrentUid();
   axios.get(`${baseUrl}/boards.json?orderBy="uid"&equalTo="${currentUid}"`)
@@ -20,4 +22,4 @@ const getAllBoards = () => new Promise((resolve, reject) => {
     }).catch((err) => reject(err));
 });
 
-export default { getAllBoards };
+export default { getAllBoards, deleteBoard };
